sumeera_task/requirements.txt
requests==2.32.3
pandas==2.2.3
    # via
    #   -r requirements.in

psycopg2-binary==2.9.9
    # via
    #   -r requirements.in

openpyxl==3.1.2

boto3==1.34.34

attrs==25.1.0
    # via
    #   -r requirements.in
    #   outcome
    #   trio
certifi==2025.1.31
    # via
    #   -r requirements.in
    #   selenium
cffi==1.17.1
    # via
    #   -r requirements.in
    #   cryptography
cryptography==43.0.3
    # via
    #   -r requirements.in
    #   pyopenssl
cssselect==1.2.0
    # via
    #   -r requirements.in
    #   parsel
exceptiongroup==1.2.2
    # via
    #   trio
    #   trio-websocket
h11==0.14.0
    # via
    #   -r requirements.in
    #   wsproto
idna==3.10
    # via
    #   -r requirements.in
    #   trio
jmespath==1.0.1
    # via
    #   -r requirements.in
    #   parsel
lxml==5.3.1
    # via
    #   lxml-html-clean
    #   parsel
lxml-html-clean==0.4.1
    # via -r requirements.in
outcome==1.3.0.post0
    # via
    #   -r requirements.in
    #   trio
    #   trio-websocket
packaging==24.2
    # via
    #   -r requirements.in
    #   parsel
parsel==1.10.0
    # via -r requirements.in
pycparser==2.22
    # via
    #   -r requirements.in
    #   cffi
pymysql==1.1.1
    # via -r requirements.in
pyopenssl==25.0.0
    # via -r requirements.in
pysocks==1.7.1
    # via
    #   -r requirements.in
    #   urllib3
selenium==4.29.0
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   -r requirements.in
    #   trio
sortedcontainers==2.4.0
    # via
    #   -r requirements.in
    #   trio
trio==0.29.0
    # via
    #   -r requirements.in
    #   selenium
    #   trio-websocket
trio-websocket==0.12.2
    # via
    #   -r requirements.in
    #   selenium
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   pyopenssl
    #   selenium
urllib3[socks]==2.3.0
    # via
    #   -r requirements.in
    #   selenium
w3lib==2.3.1
    # via
    #   -r requirements.in
    #   parsel
websocket-client==1.8.0
    # via
    #   -r requirements.in
    #   selenium
wsproto==1.2.0
    # via
    #   -r requirements.in
    #   trio-websocket
mysql-connector-python==8.0.29

webdriver-manager==4.0.1